requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.12.0
cycler==0.12.1
dateparser==1.2.2
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.6.3
multitasking==0.0.11
numpy==2.3.1
nvidia-nccl-cu12==2.27.6
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
pycryptodome==3.23.0
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2024.11.6
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
soupsieve==2.7
ta==0.11.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
xgboost==3.0.2
yarl==1.20.1
yfinance==0.2.65
